Fix math and notes
The most important problem was backslashes that were not escaped.

In the BRSA code, I also changed eta to epsilon to match the paper.

I also replaced a couple of duplicate equations with references.

For notes, the problem was that the `note` directive was mistaken for
another attribute or parameter. It needs to be separated by at least two
blank lines to be recognized as a directive. Instead I replaced it with
a Note section, which looks better in the raw docstring and looks the
same in HTML.

Fixes issue #134.