docs: Use sphinx-gallery
This PR only introduces the infrastructure and an eventseg example.
The rest of the examples will need to be converted in future PRs.

Note that the README files have to have the "txt" extension and the
RST titles within have to be formatted hierarchically at least for the
first two levels (`examples` and its subdirectories) so they are
displayed properly in the resulting HTML.

Fixes #268.