eventseg: Fix numerical issue and add utility function (#292)
Changed the transition probability to scale with the number of timepoints, which should not change the solution (all valid paths have the same number of transitions) but gives better numerical stability. Also, added a utility function for computing weighted variance, which will be useful for the notebooks I'll be releasing in the future for the Schema project.