dev: Fix failing tests (#305)
Fix Theano BLAS warning.

Also, reinstall setuptools in Conda env. This addresses errors such as the one in seen PR #305.

If the PyPI version of setuptools is newer than the one in Conda and
backwards incompatible, the build process can break.

When used with the `-I` flag, `pip install` will overwrite the Conda
`pkgs` cache, which is used for creating new environments.