dev: Drop support for Python 3.4, add 3.6 testing
Mypy dropped support for Python 3.4 in version 0.700. Technically, we
could separate the Python version in developer requirements from the
Python version in user requirements (where Mypy is not a dependency),
but I think that would complicate our setup too much.

Also switch to Ubuntu 16.04 for Travis testing, since support for 14.04
ends in April 2019.