Added example notebook tests
Notebooks have been added under docs/examples. Examples that need
to fetch data before running can include this in a download_data.sh
script that is present in their working directory. When tests are
run in CI environment on della, these scripts will be invoked and
data will be fetched into a cache if it isn't already present.
The data will then be copied to the working directory of the
notebook. Code for handling this has been added to pr-check.sh
script and only runs when on della host.