change dtype=np.int to dtype=int (#526)
In [release 1.20 of
numpy](https://numpy.org/devdocs/release/1.20.0-notes.html#deprecations),
the use of `np.int` was depreciated. Users with numpy>1.19 in their
virtual environments will encounter errors due to a few usage of np.int.
Per the suggestion, a default way is to change to `dtype=int` from
`dtype=np.int`, although what exact type of int will be used depends on
the system. However, since most of the experiments won't have subject
number more than 2^32 and I suppose it is unlikely that a system
defaults to int8. I think leaving it to `dtype=int` should be fine

Co-authored-by: lcnature <lcnature@github.com>