Make voxel selection tests deterministic.
The test module test_voxel_selection had a global module
level RandomState object. It used this global in both
tests in the module. This leads to different test results
if the tests are run separately or in different order. To fix
this I moved the state into each function. This required
changing the expected values for the second test.