Support specifying pool_size on SearchLight construction
This is needed to allow MVPAVoxelSelector to run searchlight
on poolsize of one. This is because user doesn't invoke
to SearchLight.run method (that has pool_size argugment) since
it occurs in MVPAVoxelSelector.run. This way, they can specify
pool_size on the searchlight when its constructed and before
passing to MVPAVoxelSelector, allowing them to control pool_size.

This is all in an effort to clean up tests that run MPI in conjunction
with pool_size > 1. This causes issues with a lot of MPI implementations
and was causing intermittent deadlocking of tests on windows and mac.