Fixes for numpy 2 np.linalg.solve changes.
In numpy version 2.0: np.linalg.solve, The b array is only treated as a shape (M,) column vector if it is exactly 1-dimensional. In all other instances it is treated as a stack of (M, K) matrices. Previously b would be treated as a stack of (M,) vectors if b.ndim was equal to a.ndim - 1.